Development

+++ src/components/Cars/Edit.vue
@@ -20,7 +20,9 @@
                               aria-describedby="make-input-live-feedback"
                               data-vv-as="Make">
                          </b-form-input>
-                         <b-form-invalid-feedback id="make-input-live-feedback">{{ veeErrors.first('make-input') }}</b-form-invalid-feedback>
+                         <b-form-invalid-feedback id="make-input-live-feedback">
+                              {{ veeErrors.first('make-input') }}
+                         </b-form-invalid-feedback>
                     </b-form-group>
 
                     <b-form-group class="col-sm-4 mb-3" label="Model">
@@ -123,12 +125,14 @@
                     <b-form-group class="col-sm-4 mb-3" label="Engine displacement">
                          <b-form-input placeholder="3.0" name="displacement-input"
                               v-model="car.displacement"
-                              v-validate="{ required: false, decimal:'2',min_value:0.1 }"
+                              v-validate="{ required: false, decimal:'1',min_value:0.1 }"
                               :state="validateState('displacement-input')" 
                               aria-describedby="displacement-input-live-feedback"
                               data-vv-as="Engine displacement">
                          </b-form-input>
-                         <b-form-invalid-feedback id="displacement-input-live-feedback">{{ veeErrors.first('displacement-input') }}</b-form-invalid-feedback>
+                         <b-form-invalid-feedback id="displacement-input-live-feedback">
+                              {{ veeErrors.first('displacement-input') }}
+                              </b-form-invalid-feedback>
 
                     </b-form-group>
                     <b-form-group class="col-sm-4 mb-3" label="Power">
@@ -267,17 +271,21 @@
                                         aria-describedby="repair-name-input-live-feedback"
                                         data-vv-as="name">
                                    </b-form-input>
-                                   <b-form-invalid-feedback id="repair-name-input-live-feedback">{{ veeErrors.first('repair-name-input') }}</b-form-invalid-feedback>
+                                   <b-form-invalid-feedback id="repair-name-input-live-feedback">
+                                        {{ veeErrors.first('repair-name-input') }}
+                                   </b-form-invalid-feedback>
                               </b-form-group>
                               <b-form-group class="col-sm-5 mb-3">
                                    <b-form-input placeholder="160" name="price-input" type="number" step=".01"
                                         v-model="repairPrice"
-                                        v-validate="{ required: true, min_value:1 }"
+                                        v-validate="{ required: true, decimal: 2, min_value:0.01 }"
                                         :state="validateState('price-input')" 
                                         aria-describedby="price-input-live-feedback"
                                         data-vv-as="price">
                                    </b-form-input>
-                                   <b-form-invalid-feedback id="price-input-live-feedback">{{ veeErrors.first('price-input') }}</b-form-invalid-feedback>
+                                   <b-form-invalid-feedback id="price-input-live-feedback">
+                                        {{ veeErrors.first('price-input') }}
+                                   </b-form-invalid-feedback>
                               </b-form-group>
                               <div class="col-sm-2 mb-3">
                                    <b-button size="sm" @click="addRepairRow()">
@@ -395,74 +403,69 @@ export default {
                               vm.equipmentCodes.push(element.code);
                          });
                     }
-                    if(response.status == 401) 
+               })
+               .catch(function (error) {
+                    vm.alertMessage = error.response.data;
+                    vm.dangerAlert = true;
+                    vm.alertFlag = true;
+                    vm.loading = false;
+                    if(error.response.status == 401) 
                     {
                          vm.$cookies.remove('token');
                          vm.$cookies.remove('user-email');
                          vm.$cookies.remove('role');
                          vm.$cookies.remove('user');
                          vm.$cookies.remove('currency');
-                         window.location.href = '/login';
-                    } 
-               })
-               .catch(function (error) {
-                    console.log(error);
-                    vm.alertMessage = error.response.data;
-                    vm.dangerAlert = true;
-                    vm.alertFlag = true;
-                    vm.loading = false;
+                         vm.$router.push('/login');
+                    }
                });               
           },
           fetchCarRepairs() {
                var vm = this;
                axios.get(backEndUrl + `/api/cars/${vm.$route.params.id}/repairs`, {
-                    headers: {
-                         Authorization: 'Bearer ' + window.$cookies.get('token')
-                    }
+                    headers: { Authorization: 'Bearer ' + window.$cookies.get('token') }
                })
                .then(function (response) {
                     if(response.status == 200)
-                         vm.repairs = response.data;
-                    if(response.status == 401) 
+                         vm.repairs = response.data;  
+               })
+               .catch(function (error) {
+                    vm.alertMessage = error.response.data;
+                    vm.dangerAlert = true;
+                    vm.alertFlag = true;
+                    if(error.response.status == 401) 
                     {
                          vm.$cookies.remove('token');
                          vm.$cookies.remove('user-email');
                          vm.$cookies.remove('role');
                          vm.$cookies.remove('user');
                          vm.$cookies.remove('currency');
-                         window.location.href = '/login';
-                    }   
-               })
-               .catch(function (error) {
-                    vm.alertMessage = error.response.data;
-                    vm.dangerAlert = true;
-                    vm.alertFlag = true;
-                    console.log(error);
+                         vm.$router.push('/login');
+                    }
                });               
           },
           updateCar() {
                let vm = this;
                axios.put(backEndUrl + `/api/cars/${this.$route.params.id}`, this.car, {
-                    headers: {
-                         Authorization: 'Bearer ' + window.$cookies.get('token')
-                    }
+                    headers: { Authorization: 'Bearer ' + window.$cookies.get('token') }
                })
-               .then(function (response) {                       
-                    if(response.status == 401) 
+               .then(function (response){
+                    if(response.status == 204 && vm.repairs.length == 0)
+                         vm.$router.push(`/cars/${vm.$route.params.id}`);
+               })
+               .catch(function (error) {
+                    vm.alertMessage = error.response.data;
+                    vm.dangerAlert = true;
+                    vm.alertFlag = true;
+                    if(error.response.status == 401) 
                     {
                          vm.$cookies.remove('token');
                          vm.$cookies.remove('user-email');
                          vm.$cookies.remove('role');
                          vm.$cookies.remove('user');
                          vm.$cookies.remove('currency');
-                         window.location.href = '/login';
-                    } 
-               })
-               .catch(function (error) {
-                    vm.alertMessage = error.response.data;
-                    vm.dangerAlert = true;
-                    vm.alertFlag = true;
-                    console.log(error);
+                         vm.$router.push('/login');
+                    }
                });     
           },
           validateState(ref) {
@@ -490,7 +493,10 @@ export default {
                     this.alertFlag = true;
                     console.log(this.car.base64images);
                     this.updateAll();
-               }            
+               }  
+               else
+                    this.$el.querySelector('[name="' + 
+                         this.$validator.errors.items[0].field + '"]').scrollIntoView(false);
           },
           deleteEquipmentRow (ind) {
                this.car.equipment.splice(ind, 1);
@@ -549,31 +555,31 @@ export default {
                     this.$validator.reset('price-input');
                     this.$validator.reset('repair-name-input');
                }
+               else
+                    this.$el.querySelector('[name="' + 
+                         this.$validator.errors.items[0].field + '"]').scrollIntoView(false);
           },
           getImage(image, saveTo){       
                axios.post(backEndUrl + "/api/get-image2", image, {
-                    headers: {
-                         Authorization: 'Bearer ' + window.$cookies.get('token')
-                    }
+                    headers: {Authorization: 'Bearer ' + window.$cookies.get('token')}
                })
                .then(function (response) {
                     if(response.status == 200)
-                         saveTo.push(response.data);               
-                    if(response.status == 401) 
+                         saveTo.push(response.data);                  
+               })
+               .catch(function (error) {
+                    this.alertMessage = error.response.data;
+                    this.dangerAlert = true;
+                    this.alertFlag = true;
+                    if(error.response.status == 401) 
                     {
                          this.$cookies.remove('token');
                          this.$cookies.remove('user-email');
                          this.$cookies.remove('role');
                          this.$cookies.remove('user');
                          this.$cookies.remove('currency');
-                         window.location.href = '/login';
-                    }    
-               })
-               .catch(function (error) {
-                    this.alertMessage = error.response.data;
-                    this.dangerAlert = true;
-                    this.alertFlag = true;
-                    console.log(error);
+                         this.$router.push('/login');
+                    }
                });
           },
           getImagesRecursive(n, from, to){
@@ -583,62 +589,57 @@ export default {
           deleteAllCarRepairs(){
                let vm = this;
                axios.delete(backEndUrl + `/api/cars/${vm.$route.params.id}/repairs`, {
-                    headers: {
-                         Authorization: 'Bearer ' + window.$cookies.get('token')
-                    }
+                    headers: { Authorization: 'Bearer ' + window.$cookies.get('token')}
                })
                .then(function (response) {
                     if(response.status == 200)
                          vm.insertMultipleRepairs();
-                    if(response.status == 401) 
+               })
+               .catch(function (error) {
+                    vm.alertMessage = error.response.data;
+                    vm.dangerAlert = true;
+                    vm.alertFlag = true;
+                    if(error.response.status == 401) 
                     {
                          vm.$cookies.remove('token');
                          vm.$cookies.remove('user-email');
                          vm.$cookies.remove('role');
                          vm.$cookies.remove('user');
                          vm.$cookies.remove('currency');
-                         window.location.href = '/login';
-                    } 
-               })
-               .catch(function (error) {
-                    vm.alertMessage = error.response.data;
-                    vm.dangerAlert = true;
-                    vm.alertFlag = true;
-                    console.log(error);
+                         vm.$router.push('/login');
+                    }
                })
           },
           insertMultipleRepairs(){
                let vm = this;
                axios.post(backEndUrl + `/api/cars/${vm.$route.params.id}/repairs`, vm.repairs, {
-                    headers: {
-                         Authorization: 'Bearer ' + window.$cookies.get('token')
-                    }
+                    headers: {Authorization: 'Bearer ' + window.$cookies.get('token') }
                })
                .then(function (response) {
                     if(response.status == 200)
-                    {
                          vm.$router.push(`/cars/${vm.$route.params.id}`);
-                    }
-                    if(response.status == 401) 
+               })
+               .catch(function (error) {
+                    vm.alertMessage = error.response.data;
+                    vm.dangerAlert = true;
+                    vm.alertFlag = true;
+                    if(error.response.status == 401) 
                     {
                          vm.$cookies.remove('token');
                          vm.$cookies.remove('user-email');
                          vm.$cookies.remove('role');
                          vm.$cookies.remove('user');
                          vm.$cookies.remove('currency');
-                         window.location.href = '/login';
-                    } 
-               })
-               .catch(function (error) {
-                    vm.alertMessage = error.response.data;
-                    vm.dangerAlert = true;
-                    vm.alertFlag = true;
-                    console.log(error);
+                         vm.$router.push('/login');
+                    }
                })
           },
-          updateAll(){               
-               this.deleteAllCarRepairs();
+          updateAll(){  
                this.updateCar();
+               console.log(this.repairs.length);
+               if(this.repairs.length > 0)             
+                    this.deleteAllCarRepairs();
+               
           },
           removeImageFromList(index){         
                this.car.base64images.splice(index, 1);
